TST/DEV Add MatlabTesterHelper to s/t/utils.py
Helpful for checking if scipy is installed before
running tests
